Work out SolrAndKafkaIntegrationTest
Each of the 2 tests still needs to be run separately - need to fix up running both in same run.
